Complete website landing page and improve theming
This adds additional changes to the initial
website reboot changes that introduced Sphinx
for the docs part. The landing page is finished
here to make it look more or less exact same as
in the Figma desings.

The styling for the actual docs content is quite
minimal with the basic setup that comes from the
Furo theme and a few additional color variables,
logos etc.

The landing page is styled with TailwindCSS and
it can be used to style other custom templates
as well. The Furo theme is also updated to version
`2022.06.21` that is used in other Aiven docs
as well.